+++ .idea/vcs.xml
@@ -2,5 +2,7 @@
 <project version="4">
   <component name="VcsDirectoryMappings">
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
+    <mapping directory="D:/EC327/test2" vcs="Git" />
+    <mapping directory="D:/EC327/testProject" vcs="Git" />
   </component>
 </project>
